requirements.txt
pyyaml
flask
jinja2
flask_cors
markdown
fontawesomefree